Split runtime_root daemon tests for AG12 merge-forward
