+++ crates/atm-daemon/src/tests/runtime_root/local_ipc.rs
@@ -1,0 +1,259 @@
+use super::*;
+
+#[test]
+#[serial_test::serial(env)]
+fn local_ipc_runtime_round_trips_send_after_add_member_roster_state() {
+    install_retained_runtime_factory();
+    let tempdir = TempDir::new().expect("tempdir");
+    let atm_home = tempdir.path().join("atm-home");
+    let workspace_dir = tempdir.path().join("workspace");
+    let db_path = tempdir.path().join("mail.db");
+    std::fs::create_dir_all(&atm_home).expect("atm home dir");
+    std::fs::create_dir_all(&workspace_dir).expect("workspace dir");
+    write_team_config(&atm_home, &[]);
+    write_workspace_config(&workspace_dir);
+    add_member_via_retained_admin(
+        &db_path,
+        &atm_home,
+        TEST_TEAM,
+        ROLE_TEAM_LEAD,
+        &workspace_dir,
+    );
+    add_member_via_retained_admin(&db_path, &atm_home, TEST_TEAM, "qa-a", &workspace_dir);
+
+    let _env = EnvGuard::set_many([
+        ("ATM_HOME", Some(atm_home.to_str().expect("utf8 atm home"))),
+        (
+            "ATM_CONFIG_HOME",
+            Some(tempdir.path().to_str().expect("utf8 config home")),
+        ),
+        (
+            SQLITE_RUNTIME_PATH_ENV,
+            Some(db_path.to_str().expect("utf8 sqlite db path")),
+        ),
+        ("HOME", Some(tempdir.path().to_str().expect("utf8 home"))),
+        ("USERPROFILE", None),
+    ]);
+    let socket_path = tempdir.path().join("daemon.sock");
+    let server_transport = LocalIpcServerTransportAdapter::new();
+    let runtime = server_transport
+        .prepare_runtime_at_socket_path_for_home(socket_path.clone(), &atm_home)
+        .expect("prepare runtime");
+    let mut runtime = runtime;
+    let endpoint_guard = runtime.take_endpoint_guard().expect("take endpoint guard");
+    let (lifecycle, _reset) = {
+        let lifecycle = LifecycleControlSourceAdapter::install().expect("install lifecycle");
+        let reset = LifecycleFlagResetGuard::install(lifecycle.clone());
+        (lifecycle, reset)
+    };
+    let dispatcher: Arc<dyn RequestDispatcher + Send + Sync> =
+        Arc::new(DaemonRequestDispatcher::new_for_test(
+            atm_home.clone(),
+            RuntimeStatusCache::new(),
+            db_path.clone(),
+        ));
+    let (serve_result_tx, serve_result_rx) = mpsc::channel();
+    let (ready_tx, ready_rx) = mpsc::sync_channel(1);
+
+    let join = std::thread::spawn(move || {
+        let result = runtime.serve_with_runtime_hooks(
+            dispatcher,
+            RuntimeServeHooks {
+                endpoint_guard,
+                graceful_drain_deadline: Duration::from_millis(500),
+                force_cancel_deadline: Duration::from_secs(2),
+                begin_shutdown: || Ok(()),
+                reload_runtime_view: || Ok(()),
+                finalize_shutdown: || {},
+                publish_ready: move || {
+                    ready_tx.send(()).map_err(|_| {
+                        AtmError::daemon_unavailable(
+                            "send round-trip test failed to observe the daemon ready signal",
+                        )
+                        .with_recovery(
+                            "Rerun the same-host daemon test after restoring the bounded ready-signal handshake.",
+                        )
+                    })
+                },
+            },
+        );
+        serve_result_tx.send(result).expect("send serve result");
+    });
+
+    let mut stream = connect_daemon_local_ipc_until_ready(&socket_path, ready_rx);
+    configure_test_local_ipc_timeouts(&stream);
+    let request = RequestEnvelope::Send(SendRequestEnvelope::Compose(
+        SendRequest::new(
+            atm_home.clone(),
+            workspace_dir.clone(),
+            ROLE_TEAM_LEAD.parse().expect("caller"),
+            "qa-a@test-team",
+            TEST_TEAM.parse().expect("team"),
+            SendMessageSource::Inline("hello local ipc".to_string()),
+            None,
+            false,
+            None,
+            false,
+        )
+        .expect("send request"),
+    ));
+    let request_id = next_request_id();
+    let frame = atm_core::protocol::request_to_frame_payload(request_id, request).expect("frame");
+    atm_core::protocol::write_frame(&mut stream, &frame, "write send frame").expect("write");
+    stream.flush().expect("flush");
+    let response_frame =
+        atm_core::protocol::read_frame(&mut stream, "read send frame", "send frame too large")
+            .expect("read frame")
+            .expect("response frame");
+    let (response_id, response) =
+        atm_core::protocol::response_from_frame_payload(response_frame).expect("decode response");
+    assert_eq!(response_id, request_id);
+    match response {
+        ResponseEnvelope::Send(SendResponseEnvelope::Sent(outcome)) => {
+            assert_eq!(outcome.outcome.as_str(), "sent");
+        }
+        other => panic!("unexpected response: {other:?}"),
+    }
+
+    lifecycle.set_terminate_for_test(true);
+    serve_result_rx
+        .recv_timeout(Duration::from_secs(15))
+        .expect("recv serve result")
+        .expect("serve runtime result");
+    join.join().expect("join serve thread");
+}
+
+#[test]
+#[serial_test::serial(env)]
+fn local_ipc_client_preflight_round_trips_ack_required_send_after_add_member_roster_state() {
+    install_retained_runtime_factory();
+    let tempdir = TempDir::new().expect("tempdir");
+    let atm_home = tempdir.path().join("atm-home");
+    let workspace_dir = tempdir.path().join("workspace");
+    let db_path = tempdir.path().join("mail.db");
+    std::fs::create_dir_all(&atm_home).expect("atm home dir");
+    std::fs::create_dir_all(&workspace_dir).expect("workspace dir");
+    write_team_config(&atm_home, &[]);
+    write_workspace_config(&workspace_dir);
+    add_member_via_retained_admin(
+        &db_path,
+        &atm_home,
+        TEST_TEAM,
+        ROLE_TEAM_LEAD,
+        &workspace_dir,
+    );
+    add_member_via_retained_admin(&db_path, &atm_home, TEST_TEAM, "qa-a", &workspace_dir);
+
+    let _env = EnvGuard::set_many([
+        ("ATM_HOME", Some(atm_home.to_str().expect("utf8 atm home"))),
+        (
+            "ATM_CONFIG_HOME",
+            Some(tempdir.path().to_str().expect("utf8 config home")),
+        ),
+        (
+            SQLITE_RUNTIME_PATH_ENV,
+            Some(db_path.to_str().expect("utf8 sqlite db path")),
+        ),
+        ("HOME", Some(tempdir.path().to_str().expect("utf8 home"))),
+        ("USERPROFILE", None),
+    ]);
+    let socket_path = tempdir.path().join("daemon.sock");
+    let server_transport = LocalIpcServerTransportAdapter::new();
+    let runtime = server_transport
+        .prepare_runtime_at_socket_path_for_home(socket_path.clone(), &atm_home)
+        .expect("prepare runtime");
+    let mut runtime = runtime;
+    let endpoint_guard = runtime.take_endpoint_guard().expect("take endpoint guard");
+    let (lifecycle, _reset) = {
+        let lifecycle = LifecycleControlSourceAdapter::install().expect("install lifecycle");
+        let reset = LifecycleFlagResetGuard::install(lifecycle.clone());
+        (lifecycle, reset)
+    };
+    let dispatcher: Arc<dyn RequestDispatcher + Send + Sync> =
+        Arc::new(DaemonRequestDispatcher::new_for_test(
+            atm_home.clone(),
+            RuntimeStatusCache::new(),
+            db_path.clone(),
+        ));
+    let (serve_result_tx, serve_result_rx) = mpsc::channel();
+    let (ready_tx, ready_rx) = mpsc::sync_channel(1);
+
+    let join = std::thread::spawn(move || {
+        let result = runtime.serve_with_runtime_hooks(
+            dispatcher,
+            RuntimeServeHooks {
+                endpoint_guard,
+                graceful_drain_deadline: Duration::from_millis(500),
+                force_cancel_deadline: Duration::from_secs(2),
+                begin_shutdown: || Ok(()),
+                reload_runtime_view: || Ok(()),
+                finalize_shutdown: || {},
+                publish_ready: move || {
+                    ready_tx.send(()).map_err(|_| {
+                        AtmError::daemon_unavailable(
+                            "send round-trip test failed to observe the daemon ready signal",
+                        )
+                        .with_recovery(
+                            "Rerun the same-host daemon test after restoring the bounded ready-signal handshake.",
+                        )
+                    })
+                },
+            },
+        );
+        serve_result_tx.send(result).expect("send serve result");
+    });
+
+    let _stream = connect_daemon_local_ipc_until_ready(&socket_path, ready_rx);
+    let endpoint =
+        atm_daemon_client::DaemonLocalIpcEndpoint::new(socket_path.clone()).expect("endpoint");
+    let request = SendRequest::new(
+        atm_home.clone(),
+        workspace_dir.clone(),
+        ROLE_TEAM_LEAD.parse().expect("caller"),
+        "qa-a@test-team",
+        TEST_TEAM.parse().expect("team"),
+        SendMessageSource::Inline("hello local ipc ack-required".to_string()),
+        None,
+        true,
+        None,
+        false,
+    )
+    .expect("send request");
+    let request = RequestEnvelope::Send(SendRequestEnvelope::Compose(request));
+    let envelope = atm_daemon_client::RpcEnvelope::encode_body(
+        atm_daemon_client::RpcHeader::new(
+            atm_daemon_client::RequestId::new(next_request_id().into_inner()).expect("request id"),
+            atm_daemon_client::MessageKind::SendComposeRequest,
+        ),
+        &request,
+    )
+    .expect("encode request");
+
+    let mut verified = atm_daemon_client::verify_connection_compatibility(
+        &endpoint,
+        atm_daemon_client::CompatibilityPreflight {
+            client_release: atm_daemon_client::ReleaseVersion::current(),
+            wire_version: atm_core::protocol::ATM_FRAME_VERSION_V1,
+        },
+        Duration::from_secs(3),
+    )
+    .expect("preflight compatible");
+    let response = verified
+        .dispatch_write(&endpoint, envelope, Duration::from_secs(3))
+        .expect("dispatch write");
+    let response: ResponseEnvelope = response.decode_body().expect("decode response");
+    match response {
+        ResponseEnvelope::Send(SendResponseEnvelope::Sent(outcome)) => {
+            assert_eq!(outcome.outcome.as_str(), "sent");
+            assert!(outcome.requires_ack);
+        }
+        other => panic!("unexpected response: {other:?}"),
+    }
+
+    lifecycle.set_terminate_for_test(true);
+    serve_result_rx
+        .recv_timeout(Duration::from_secs(15))
+        .expect("recv serve result")
+        .expect("serve runtime result");
+    join.join().expect("join serve thread");
+}
